WaitForCondition Step: Fix race condition
Sometimes the evaluation of the check script takes more then the 0.1 sec
between the runs, which means a second run is triggered while the first
one is still running. If the first on is now successful, and the second
one returns before the reset function is fully executed, the completion
handler might be called multiple times, advancing more steps then it
should.

To fix this, we now only trigger the javascript check function 0.1 sec
after the last one returned.